src/pyproject.toml
[project]
name = "grpc"
version = "0.6.0"
description = ""
readme = "README.md"
requires-python = ">=3.10,<3.12"
authors = [
    { name = "SD Traders", email = "torrino07@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

dependencies = [
    "fastapi>=0.115.12,<1.0.0",
    "websockets>=15.0.1",
    "uvicorn>=0.34.3,<1.0.0",
    "pydantic>=2.11.5,<3.0.0",
    "requests>=2.32.4,<3.0.0",
    "itsdangerous>=2.2.0,<3.0.0",
    "starlette>=0.40.0,<0.47.0",
    "debugpy>=1.8.14,<2.0.0",
    "psycopg>=3.2.9,<4.0.0",
    "psycopg-pool>=3.2.6,<4.0.0",
    "protobuf==4.21.12",
    "grpcio-tools==1.59.0",
    "grpcio==1.59.0",
    "boto3>=1.35.39,<2.0.0",
    "python-dotenv>=1.1.0",
    "pyzmq>=27.0.0",
    "psutil>=7.0.0"
]
